# lib/api-zod/src/generated/types/authSignInRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthSignInRequest {
  /** @maxLength 254 */
  email: string;
  /**
   * @minLength 1
   * @maxLength 128
   */
  password: string;
}