# lib/api-zod/src/generated/types/authUserResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUser } from "./publicUser";

export interface AuthUserResponse {
  user: PublicUser;
}