# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicUser {
  id: string;
  email: string;
  name: string | null;
  createdAt: Date;
}